utils: Move extension functions to their own file
This continues the effort so modularize the monolithic "Utils.kt" file.
